+++ xcode/MAME4iOS/SoftwareList.m
@@ -9,15 +9,17 @@
 #import "XmlFile.h"
 #import "ZipFile.h"
 #import "GameInfo.h"
+#import <CommonCrypto/CommonDigest.h>
 
 #define DebugLog 0
 #if DebugLog == 0 || !defined(DEBUG)
 #define NSLog(...) (void)0
 #endif
 
-#define STR(x)      ((NSString*)([x isKindOfClass:[NSString class]]     ? x : @""))
-#define DICT(x) ((NSDictionary*)([x isKindOfClass:[NSDictionary class]] ? x : @{}))
-#define LIST(x)      ((NSArray*)([x isKindOfClass:[NSArray class]]      ? x : (x ? @[x] : @[])))
+// SAFE access to values in XML files
+#define STR(x)  ({id s = x; [s isKindOfClass:[NSString class]] ? (NSString*)(s) : @"";})
+#define DICT(x) ({id d = x; [d isKindOfClass:[NSDictionary class]] ? (NSDictionary*)(d) : @{};})
+#define LIST(x) ({id a = x; [a isKindOfClass:[NSArray class]] ? (NSArray*)(a) : (a ? @[a] : @[]);})
 
 #define ZIP_FILE_TYPES    @[@"zip", @"7z"]
 
@@ -27,7 +29,7 @@
 
 @implementation SoftwareList
 {
-    NSString* hash_dir;
+    NSString* hash_zip;
     NSString* roms_dir;
     NSCache*  software_list_cache;
 }
@@ -37,7 +39,7 @@
 - (instancetype)initWithPath:(NSString*)root
 {
     self = [super init];
-    hash_dir = [root stringByAppendingPathComponent:@"hash"];
+    hash_zip = [root stringByAppendingPathComponent:@"hash.zip"];
     roms_dir = [root stringByAppendingPathComponent:@"roms"];
     software_list_cache = [[NSCache alloc] init];
     return self;
@@ -51,12 +53,11 @@
         NSArray* list = [software_list_cache objectForKey:@"*"];
         if (list == nil) {
             @autoreleasepool {
-                list = [NSFileManager.defaultManager contentsOfDirectoryAtPath:hash_dir error:nil];
-                list = [list valueForKey:@"stringByDeletingPathExtension"];
+                list = [getZipFileNames(hash_zip) valueForKeyPath:@"lastPathComponent.stringByDeletingPathExtension.lowercaseString"];
                 [software_list_cache setObject:list forKey:@"*"];
             }
         }
-        return list;
+        return list; 
     }
 }
 
@@ -75,10 +76,10 @@
     }
 }
 
-// get games for a system list can be a comma separated list of softlist names
+// get games for a system
 - (NSArray<NSDictionary*>*)getGamesForSystem:(NSDictionary*)system {
 
-    NSString* lists = system.gameSoftware;
+    NSString* lists = system.gameSoftware; // this can be a comma separated list of softlist names
     
     NSParameterAssert(![lists containsString:@" "]);
     if (lists.length == 0)
@@ -104,18 +105,90 @@
     return games;
 }
 
-// install a XML or ZIP file
-- (BOOL)installFile:(NSString*)path {
- 
-    @autoreleasepool {
-        if ([ZIP_FILE_TYPES containsObject:path.pathExtension.lowercaseString])
-            return [self installZIP:path];
+// get name of software list for a romset, this is used at install time.
+//
+// this is to solve the "pacman problem", there is a pacman.zip for almost every system
+// given a zip file we need to find out what software list directory to copy it to, and this
+// involves searching *all* the software list XMLs looking for a match. we do some cacheing/etc
+// to make this not terribly slow, but we still need to search
+//
+// we use a cached pre-computed database to find the possible software list names to search
+//
+- (nullable NSString*)getSoftwareListNameForRomset:(NSString*)path named:(NSString*)name {
+    
+    if (name.length == 0)
+        name = path.lastPathComponent.stringByDeletingPathExtension.lowercaseString;
+    
+    // get subset of software lists to search
+    NSDictionary* soft_list_db = [self getSoftwareListDatabase];
+    NSArray* list_names = LIST(soft_list_db[name]);
+    if (list_names.count == 0)
+        return nil;
 
-        if ([@[@"xml"] containsObject:path.pathExtension.lowercaseString])
-            return [self installXML:path];
+    // get SHA1 of all files in this romset
+    NSArray* hashes = getZipFileSHA1(path);
+    if (hashes.count == 0)
+        return nil;
+
+    NSLog(@"SEARCHING SOFTWARE LISTS(%@) FOR: %@", [list_names componentsJoinedByString:@", "], name);
+
+    // figure out if this zip file is a SOFTWARE romset, by looking for it in *all* the installed software lists
+    for (NSString* list_name in list_names) @autoreleasepool {
+        
+        NSLog(@"    SEARCHING LIST(%@) FOR: %@", list_name, name);
+
+        NSDictionary* dict = [self getSoftwareListXML:list_name];
+
+        // walk all software in this list looking for a name match, then if names match check ROMs and DISKs
+        for (NSDictionary* software in LIST([dict valueForKeyPath:@"softwarelist.software"])) {
+            
+            if (![software isKindOfClass:[NSDictionary class]])
+                continue;
+
+            // check this software if the name matches
+            if ([name isEqualToString:STR(software[kSoftwareListName]).lowercaseString]) {
+                NSLog(@"        FOUND IN %@.%@ (checking ROMs)", list_name, STR(software[kSoftwareListName]));
+
+                // now make sure all the roms (and disks) for this software are in this ZIP.
+                int rom_count = 0;
+                int zip_count = 0;
+                for (NSDictionary* part in LIST(software[@"part"])) {
+                    
+                    // look at all the ROMs and see if they are in the ZIP
+                    for (NSDictionary* area in LIST([part valueForKeyPath:@"dataarea"])) {
+                        for (NSString* hash in LIST([area valueForKeyPath:@"rom.sha1"])) {
+                            if (![hash isKindOfClass:[NSString class]] || hash.length == 0)
+                                continue;
+                            NSLog(@"            ROM:%@ %@", hash, [hashes containsObject:hash.lowercaseString] ? @"FOUND" : @"**NOT** FOUND");
+                            rom_count++;
+                            if ([hashes containsObject:hash.lowercaseString])
+                                zip_count++;
+                        }
+                    }
+                    
+                    // look at all the DISKs and see if they are in the ZIP
+                    for (NSDictionary* area in LIST([part valueForKeyPath:@"diskarea"])) {
+                        for (NSString* hash in LIST([area valueForKeyPath:@"disk.sha1"])) {
+                            if (![hash isKindOfClass:[NSString class]] || hash.length == 0)
+                                continue;
+                            NSLog(@"           DISK:%@ %@", hash, [hashes containsObject:hash.lowercaseString] ? @"FOUND" : @"**NOT** FOUND");
+                            rom_count++;
+                            if ([hashes containsObject:hash.lowercaseString])
+                                zip_count++;
+                        }
+                    }
+                }
+                
+                if (rom_count != 0 && rom_count == zip_count) {
+                    NSLog(@"FOUND %@ in LIST: %@", name, list_name);
+                    return list_name;
+                }
+            }
+        }
     }
-
-    return FALSE;
+    
+    NSLog(@"DID NOT FIND %@ in *any* SOFTWARE LIST", name);
+    return nil;
 }
 
 // if this a merged romset, extract clones as empty zip files so they show up as Available
@@ -133,7 +206,7 @@
     NSLog(@"EXTRACT CLONES: %@", path);
     
     // get directory names of all files in this romset
-    NSSet* clones = [NSSet setWithArray:[getZipFiles(path) valueForKeyPath:@"stringByDeletingLastPathComponent.lowercaseString"]];
+    NSSet* clones = [NSSet setWithArray:[getZipFileNames(path) valueForKeyPath:@"stringByDeletingLastPathComponent.lowercaseString"]];
     
     if (clones.count < 2 || ![clones containsObject:@""]) {
         NSLog(@"....NOT A MERGED ROMSET");
@@ -162,54 +235,10 @@
     }
 }
 
-// delete all software
-- (void)reset {
-    for (NSString* name in [self getSoftwareListNames])  {
-        NSString* roms_path = [roms_dir stringByAppendingPathComponent:name];
-        [NSFileManager.defaultManager removeItemAtPath:roms_path error:nil];
-        NSString* titles_path = [roms_dir stringByAppendingPathComponent:[NSString stringWithFormat:@"../titles/%@", name]];
-        [NSFileManager.defaultManager removeItemAtPath:titles_path error:nil];
-    }
-    [NSFileManager.defaultManager removeItemAtPath:hash_dir error:nil];
-    [NSFileManager.defaultManager createDirectoryAtPath:hash_dir withIntermediateDirectories:NO attributes:nil error:nil];
-    [self reload];
-}
-
 #pragma mark Private methods
 
-// install a XML file
-- (BOOL)installXML:(NSString*)path {
-    
-    // load the XML file
-    NSDictionary* dict = [XmlFile dictionaryWithPath:path error:nil];
-
-    NSString* name = STR([dict valueForKeyPath:@"softwarelist.name"]);
-    NSString* desc = STR([dict valueForKeyPath:@"softwarelist.description"]);
-    NSArray*  list = LIST([dict valueForKeyPath:@"softwarelist.software"]);
-
-    if (name.length == 0 || desc.length == 0 || list.count == 0) {
-        NSLog(@"INVALID SOFTWARE LIST: %@", path);
-        return FALSE;
-    }
-    
-    NSLog(@"INSTALL SOFTWARE LIST: %@ \"%@\" (%d items)", name, desc, (int)list.count);
-    
-    NSString* hash_path = [[hash_dir stringByAppendingPathComponent:name] stringByAppendingPathExtension:@"xml"];
-    NSString* roms_path = [roms_dir stringByAppendingPathComponent:name];
-
-    // move software list xml to the `hash` directory (overwrite file)
-    [NSFileManager.defaultManager removeItemAtPath:hash_path error:nil];
-    [NSFileManager.defaultManager moveItemAtPath:path toPath:hash_path error:nil];
-    
-    // create (if needed) directory to hold software ROMs
-    [NSFileManager.defaultManager createDirectoryAtPath:roms_path withIntermediateDirectories:NO attributes:nil error:nil];
-    
-    return TRUE;
-}
-
 // get the names of all files in a ZIP, excluding hidden files and directories.
-// TODO: this does not work for 7z files
-static NSArray<NSString*>* getZipFiles(NSString* path) {
+static NSArray<NSString*>* getZipFileNames(NSString* path) {
     NSMutableArray* files = [[NSMutableArray alloc] init];
     [ZipFile enumerate:path withOptions:ZipFileEnumFiles usingBlock:^(ZipFileInfo* info) {
         [files addObject:info.name];
@@ -217,102 +246,82 @@ static NSArray<NSString*>* getZipFiles(NSString* path) {
     return [files copy];
 }
 
-// install a ZIP file
-- (BOOL)installZIP:(NSString*)path {
+static NSString* sha1(NSData* data) {
+    if (data.length == 0)
+        return @"";
+    uint8_t digest[CC_SHA1_DIGEST_LENGTH];
+    CC_SHA1(data.bytes, (CC_LONG)data.length, digest);
 
-    // if we dont have any software lists, bail
-    NSArray* list_names = [self getSoftwareListNames];
-    if (list_names.count == 0)
-        return FALSE;
-    
-    // get names of all files in this romset, if empty zip get out.
-    NSArray* files = [getZipFiles(path) valueForKeyPath:@"lastPathComponent.lowercaseString"];
-    if (files.count == 0)
-        return FALSE;
-
-    NSString* name = path.lastPathComponent.stringByDeletingPathExtension.lowercaseString;
-    
-    // if the ROM had a name like "foobar 1.zip", "foobar (1).zip" use only the first word as the ROM name.
-    // this most likley came when a user downloaded the zip and a foobar.zip already existed, MAME ROMs are <=20 char and no spaces.
-    NSArray* words = [name componentsSeparatedByCharactersInSet:[NSCharacterSet characterSetWithCharactersInString:@" -"]];
-    if (words.count == 2 && [words.lastObject stringByTrimmingCharactersInSet:[NSCharacterSet punctuationCharacterSet]].intValue != 0)
-        name = words.firstObject;
-    
-    NSLog(@"SEARCHING SOFTWARE LISTS FOR: %@", name);
-
-    // figure out if this zip file is a SOFTWARE romset, buy looking for it in *all*
-    // ...the installed software lists, and if it is copy it to the right subdir of `roms`
-    for (NSString* list_name in list_names) @autoreleasepool {
-        
-        NSLog(@"    SEARCHING LIST(%@) FOR: %@", list_name, name);
-
-        NSDictionary* dict = [self loadSoftwareListXML:list_name cache:YES];
-
-        // walk all software in this list looking for a name match, then if names match check names of ROMs
-        for (NSDictionary* software in LIST([dict valueForKeyPath:@"softwarelist.software"])) {
-            
-            if (![software isKindOfClass:[NSDictionary class]])
-                continue;
-
-            // check this software if the name or the parent match
-            if ([name isEqualToString:STR(software[kSoftwareListName]).lowercaseString] || [name isEqualToString:STR(software[kSoftwareListParent]).lowercaseString]) {
-                NSLog(@"        FOUND IN %@.%@ (checking ROMs)", list_name, STR(software[kSoftwareListName]));
-
-                // now make sure the roms for this software are in this ZIP.
-                BOOL found_all_roms = TRUE;
-                for (NSDictionary* part in LIST(software[@"part"])) {
-                    for (NSDictionary* area in LIST([part valueForKeyPath:@"dataarea"])) {
-                        for (NSString* name in LIST([area valueForKeyPath:@"rom.name"])) {
-                            NSLog(@"            ROM:\"%@\" %@", name, [files containsObject:name.lowercaseString] ? @"FOUND" : @"**NOT** FOUND");
-                            if (![files containsObject:name.lowercaseString])
-                                found_all_roms = FALSE;
-                        }
-                    }
-                }
-                
-                if (found_all_roms) {
-                    NSLog(@"FOUND %@ in LIST: %@", name, list_name);
-                    NSString* dest_path = [[[roms_dir stringByAppendingPathComponent:list_name] stringByAppendingPathComponent:name] stringByAppendingPathExtension:path.pathExtension];
-                    
-                    // move romset to roms sub directory and overwrite any other file that is there.
-                    [NSFileManager.defaultManager removeItemAtPath:dest_path error:nil];
-                    [NSFileManager.defaultManager moveItemAtPath:path toPath:dest_path error:nil];
-                    
-                    [self extractClones:dest_path];
-                    return TRUE;
-                }
-            }
-        }
-    }
-    
-    NSLog(@"DID NOT FIND %@ in *any* SOFTWARE LIST", name);
-    return FALSE;
+    return [NSString stringWithFormat: @"%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x%02x",
+            digest[0], digest[1], digest[2], digest[3],
+            digest[4], digest[5], digest[6], digest[7],
+            digest[8], digest[9], digest[10], digest[11],
+            digest[12],digest[13], digest[14], digest[15],
+            digest[16],digest[17], digest[18], digest[19]];
 }
 
-// load software list xml file (and maybe cache it)
-- (NSDictionary*)loadSoftwareListXML:(NSString*)name cache:(BOOL)cache {
+// get the SHA1 of all files in a ZIP, excluding hidden files and directories.
+// TODO: this does not work for 7z files
+static NSArray<NSString*>* getZipFileSHA1(NSString* path) {
+    NSMutableArray* hashes = [[NSMutableArray alloc] init];
+    [ZipFile enumerate:path withOptions:(ZipFileEnumFiles | ZipFileEnumLoadData) usingBlock:^(ZipFileInfo* info) {
+        [hashes addObject:sha1(info.data)];
+    }];
+    return [hashes copy];
+}
+
+// load software list xml file, but check cache first
+- (NSDictionary*)getSoftwareListXML:(NSString*)name {
     NSParameterAssert(name.length != 0);
-    NSString* path = [[hash_dir stringByAppendingPathComponent:name] stringByAppendingPathExtension:@"xml"];
+    NSString* key = [name stringByAppendingPathExtension:@"xml"];   // use swlist.xml as key to not conflict with getSoftwareList
     
     @synchronized (self) {
-        NSDictionary* dict = [software_list_cache objectForKey:path];
+        NSDictionary* dict = [software_list_cache objectForKey:key];
         if (dict == nil) {
             @autoreleasepool {
-                dict = [XmlFile dictionaryWithPath:path error:nil];
-                if (cache)
-                    [software_list_cache setObject:dict forKey:path];
+                dict = [self loadSoftwareListXML:name];
+                [software_list_cache setObject:dict forKey:key];
             }
         }
         return dict;
     }
 }
 
-// load software list, and validate the ROMs
+// load software list xml file
+// NOTE all the software lists are inside `hash.zip` (not as separate files)
+- (NSDictionary*)loadSoftwareListXML:(NSString*)name {
+    
+    __block NSDictionary* dict = nil;
+    [ZipFile enumerate:hash_zip withOptions:ZipFileEnumFiles usingBlock:^(ZipFileInfo* info) {
+        
+        if (dict != nil)
+            return;
+        
+        if (![info.name.pathExtension.lowercaseString isEqualToString:@"xml"])
+            return;
+        
+        if (![info.name.lastPathComponent.stringByDeletingPathExtension.lowercaseString isEqualToString:name.lowercaseString])
+            return;
+
+        if (info.data == nil)
+            return;
+        
+        dict = [XmlFile dictionaryWithData:info.data error:nil];
+    }];
+
+    return dict;
+}
+
+// load software list, and validate the ROMs (ie filter software list down to only Available software)
 - (NSArray*)loadSoftwareList:(NSString*)list_name {
     NSParameterAssert(list_name.length != 0);
     
-    // load the XML file
-    NSDictionary* dict = [self loadSoftwareListXML:list_name cache:FALSE];
+#if defined(DEBUG) && DebugLog != 0
+    NSTimeInterval time = [NSDate timeIntervalSinceReferenceDate];
+#endif
+
+    // load the XML file, NOTE we dont need to use cached XML because our result gets cached.
+    NSDictionary* dict = [self loadSoftwareListXML:list_name];
     
     // get all the software in the list
     NSArray* list = LIST([dict valueForKeyPath:@"softwarelist.software"]);
@@ -328,8 +337,6 @@ static NSArray<NSString*>* getZipFiles(NSString* path) {
         if (soft_name.length == 0)
             return FALSE;
         
-        // TODO: just checkin for zip file may not be enough if the Parent and Child ROMs are merged
-        // TODO: maybe we should add all Clones if the Parent romset is present?
         NSString* soft_path = [soft_dir stringByAppendingPathComponent:soft_name];
         for (NSString* ext in ZIP_FILE_TYPES) {
             if ([NSFileManager.defaultManager fileExistsAtPath:[soft_path stringByAppendingPathExtension:ext]])
@@ -338,17 +345,90 @@ static NSArray<NSString*>* getZipFiles(NSString* path) {
 
         return FALSE;
     }]];
-
+    
 #if defined(DEBUG) && DebugLog != 0
-    if (list.count != 0) {
-        NSLog(@"LOADING SOFTWARE LIST: %@ \"%@\"", [dict valueForKeyPath:@"softwarelist.name"], [dict valueForKeyPath:@"softwarelist.description"]);
-        for (NSDictionary* software in list)
-            NSLog(@"    %@, %@, %@, \"%@\"", software[kSoftwareListName], software[kSoftwareListYear], software[kSoftwareListPublisher], software[kSoftwareListDescription]);
-    }
+    NSLog(@"LOADING SOFTWARE LIST: %@ \"%@\" %d items, %0.3fsec", [dict valueForKeyPath:@"softwarelist.name"], [dict valueForKeyPath:@"softwarelist.description"], (int)list.count, [NSDate timeIntervalSinceReferenceDate] - time);
+    for (NSDictionary* software in list)
+        NSLog(@"    %@, %@, %@, \"%@\"", software[kSoftwareListName], software[kSoftwareListYear], software[kSoftwareListPublisher], software[kSoftwareListDescription]);
 #endif
     
     return list;
 }
 
+// get database that maps romset name to possible software lists
+- (NSDictionary*)getSoftwareListDatabase {
+    NSString* key = @"hash.dat";
+    
+    @synchronized (self) {
+        NSDictionary* dict = [software_list_cache objectForKey:key];
+        if (dict == nil) {
+            @autoreleasepool {
+                dict = [self loadSoftwareListDatabase];
+                [software_list_cache setObject:dict forKey:key];
+            }
+        }
+        return dict;
+    }
+}
+
+// load/create software list database
+- (NSDictionary*)loadSoftwareListDatabase {
+    
+    NSString* hash_dat = [hash_zip.stringByDeletingPathExtension stringByAppendingPathExtension:@"dat"];
+    
+    NSDate* zip_date = [[NSFileManager.defaultManager attributesOfItemAtPath:hash_zip error:nil] fileModificationDate];
+    NSDate* dat_date = [[NSFileManager.defaultManager attributesOfItemAtPath:hash_dat error:nil] fileModificationDate] ?: NSDate.distantPast;
+    
+    // load `HASH.DAT` if it is not older than `HASH.ZIP` otherwise build it.
+    if ([zip_date compare:dat_date] != NSOrderedDescending) {
+        NSData* data = [NSData dataWithContentsOfFile:hash_dat] ?: [[NSData alloc] init];
+        NSDictionary* dict = [NSPropertyListSerialization propertyListWithData:data options:NSPropertyListImmutable format:nil error:nil];
+        NSParameterAssert([dict isKindOfClass:[NSDictionary class]]);
+
+        if ([dict isKindOfClass:[NSDictionary class]])
+            return dict;
+    }
+    
+    NSMutableDictionary* soft_list_db = [[NSMutableDictionary alloc] init];
+    
+    // walk *all* of the software lists and build up database
+    [ZipFile enumerate:hash_zip withOptions:ZipFileEnumFiles usingBlock:^(ZipFileInfo* info) {
+        
+        if (![info.name.pathExtension.lowercaseString isEqualToString:@"xml"] || info.data == nil)
+            return;
+
+        NSDictionary* dict = [XmlFile dictionaryWithData:info.data error:nil];
+        
+        NSString* list_name = info.name.lastPathComponent.stringByDeletingPathExtension.lowercaseString;
+
+        // walk all software in this list and add to master db
+        for (NSDictionary* software in LIST([dict valueForKeyPath:@"softwarelist.software"])) {
+
+            if (![software isKindOfClass:[NSDictionary class]] || software[kSoftwareListName] == nil)
+                continue;
+            
+            NSString* name = STR(software[kSoftwareListName]).lowercaseString;
+            id val = soft_list_db[name];
+
+            if (val == nil)
+                soft_list_db[name] = list_name;
+            else
+                soft_list_db[name] = [LIST(val) arrayByAddingObject:list_name];
+        }
+    }];
+    
+    // SAVE the database to `HASH.DAT` for next time.
+    NSData* data = [NSPropertyListSerialization dataWithPropertyList:soft_list_db format:NSPropertyListBinaryFormat_v1_0 options:0 error:nil];
+    NSParameterAssert(data != nil);
+    [data writeToFile:hash_dat atomically:NO];
+    
+#if defined(DEBUG) && DebugLog != 0
+    for (NSString* key in soft_list_db.allKeys)
+        NSLog(@"    %@: %@", key, [LIST(soft_list_db[key]) componentsJoinedByString:@", "]);
+    NSLog(@"SoftwareListDatabase: %d items, HASH.DAT size=%d", (int)soft_list_db.allKeys.count, (int)[data length]);
+#endif
+    
+    return soft_list_db;
+}
 
 @end